mailcheck_nouser_back/api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysqlclient
alembic
backports.ssl
pyOpenSSL
pyzmail36
datetime
imapclient